# swagger.yaml
openapi: 3.0.3
info:
  title: CargoFlow Logistics API
  version: 2.0.0
  description: >
    API documentation for Shipment Management System.

    USER STORIES & API MAPPING:

    Auth:
    - Register → Create account + Email notification
    - Login → Access system
    - Profile → View/update user details

    Admin:
    - Dashboard → Monitor system
    - Shipments → Create & assign shipments
    - Agents → Manage delivery agents
    - Clients → Manage business clients
    - Tracking → Monitor agent locations

    Client:
    - Shipments → Create & view shipments
    - Tracking → Track deliveries
    - Business → Manage business profile

    Agent:
    - Dashboard → View assigned shipments
    - Status → Update shipment status
    - Location → Update live location
    - Duty → Update availability

    Chatbot:
    - Chat → Ask queries
    - Rebuild → Refresh knowledge base
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/v1/client/dashboard:
    get:
      tags: [Client]
      summary: Client Dashboard
      description: "User Story: As a client, I want to view my dashboard so that I can track my shipment activities."
      security:
        - bearerAuth: []
      responses:
        "200": { description: Dashboard }
  /api/v1/client/shipments:
    get:
      tags: [Client]
      summary: View Shipments
      description: "User Story: As a client, I want to view all my shipments so that I can monitor delivery status."
      security:
        - bearerAuth: []
      responses:
        "200": { description: List }

    post:
      tags: [Client]
      summary: Create Shipment
      description: "User Story: As a client, I want to create shipments so that I can send packages and receive confirmation emails."
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientShipmentCreate'
      responses:
        "201": { description: Created }
  /api/v1/client/track/{tracking_id}:
    get:
      tags: [Client]
      summary: Track Shipment
      description: "User Story: As a user, I want to track shipments using a tracking ID so that I can know delivery status without logging in."
      parameters:
        - name: tracking_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Tracking }
  /api/v1/client/business:
    post:
      tags: [Client]
      summary: Create Business
      description: "User Story: As a client, I want to create my business profile so that I can start using logistics services."
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessCreate'
      responses:
        "201": { description: Created }

    put:
      tags: [Client]
      summary: Update Business
      description: "User Story: As a client, I want to update my business profile so that my business details remain accurate."
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessCreate'
      responses:
        "200": { description: Updated }

  # ---------------- AGENT ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    AdminShipmentCreate:
      type: object
      required: [sender_id, receiver_name, receiver_phone, receiver_email, weight, price]
      properties:
        sender_id: { type: integer }
        receiver_name: { type: string }
        receiver_phone: { type: string }
        receiver_email: { type: string, format: email }
        pickup_line1: { type: string }
        pickup_city: { type: string }
        pickup_state: { type: string }
        pickup_pincode: { type: string }
        delivery_line1: { type: string }
        delivery_city: { type: string }
        delivery_state: { type: string }
        delivery_pincode: { type: string }
        pickup_lat: { type: number }
        pickup_lng: { type: number }
        delivery_lat: { type: number }
        delivery_lng: { type: number }
        weight: { type: number }
        price: { type: number }
    ClientShipmentCreate:
      allOf:
        - $ref: '#/components/schemas/AdminShipmentCreate'
    BusinessCreate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
...